requirements.txt
streamlit
PyPDF2
python-dotenv
langchain
sentence-transformers
uvicorn
ctransformers
python-box
replicate